pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0", "hatch-vcs>=0.4.0"]
build-backend = "hatchling.build"

[project]
name = "burf"
dynamic = ["version"]
description = "A tool to surf buckets on gcs"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Mohammad Razeghi", email = "razeghi71@gmail.com" }]
dependencies = [
  "google-cloud-storage==2.10.0",
  "textual==0.28.1",
  "boto3>=1.28.0",
]
keywords = ["gcs", "google-cloud-storage", "tui", "textual"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/razeghi71/burf"
Repository = "https://github.com/razeghi71/burf"

[project.scripts]
burf = "burf.burf:main"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["burf"]

